Move main out of src
