+++ main.py
@@ -1,6 +1,6 @@
 import pygame
 import sys
-from game import Game
+from src.game import Game
 
 def main():
     pygame.init()
